Separate metadata-clients.go into respective files
