+++ core/clients/metadata/device.go
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright 2017 Dell Inc.
+ * Copyright 2018 Dell Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
  * in compliance with the License. You may obtain a copy of the License at
@@ -18,7 +18,6 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"io/ioutil"
 	"net/http"
 	"net/url"
 	"strconv"
@@ -27,23 +26,6 @@ import (
 	"github.com/edgexfoundry/edgex-go/core/clients/types"
 	"github.com/edgexfoundry/edgex-go/core/domain/models"
 )
-
-var (
-	ErrResponseNil error = errors.New("Problem connecting to metadata - reponse was nil")
-	ErrNotFound    error = errors.New("Item not found")
-)
-
-/*
-Addressable client for interacting with the addressable section of metadata
-*/
-type AddressableClient interface {
-	Add(addr *models.Addressable) (string, error)
-	AddressableForName(name string) (models.Addressable, error)
-}
-
-type AddressableRestClient struct {
-	url string
-}
 
 /*
 Device client for interacting with the device section of metadata
@@ -80,54 +62,6 @@ type DeviceRestClient struct {
 }
 
 /*
-Command client for interacting with the command section of metadata
-*/
-type CommandClient interface {
-	Add(com *models.Command) (string, error)
-	Command(id string) (models.Command, error)
-	Commands() ([]models.Command, error)
-	CommandsForName(name string) ([]models.Command, error)
-	Delete(id string) error
-	Update(com models.Command) error
-}
-
-type CommandRestClient struct {
-	url string
-}
-
-/*
-Service client for interacting with the device service section of metadata
-*/
-type ServiceClient interface {
-	Add(ds *models.DeviceService) (string, error)
-	DeviceServiceForName(name string) (models.DeviceService, error)
-	UpdateLastConnected(id string, time int64) error
-	UpdateLastReported(id string, time int64) error
-}
-
-type ServiceRestClient struct {
-	url string
-}
-
-// Device Profile client for interacting with the device profile section of metadata
-type DeviceProfileClient interface {
-	Add(dp *models.DeviceProfile) (string, error)
-}
-
-type DeviceProfileRestClient struct {
-	url string
-}
-
-/*
-Return an instance of AddressableClient
-*/
-func NewAddressableClient(metaDbAddressableUrl string) AddressableClient {
-	a := AddressableRestClient{url: metaDbAddressableUrl}
-
-	return &a
-}
-
-/*
 Return an instance of DeviceClient
 */
 func NewDeviceClient(params types.EndpointParams, m clients.Endpointer) (DeviceClient, error) {
@@ -153,130 +87,6 @@ func(d *DeviceRestClient) init(params types.EndpointParams) {
 	}
 }
 
-/*
-Return an instance of CommandClient
-*/
-func NewCommandClient(metaDbCommandUrl string) CommandClient {
-	c := CommandRestClient{url: metaDbCommandUrl}
-
-	return &c
-}
-
-/*
-Return an instance of ServiceClient
-*/
-func NewServiceClient(metaDbServiceUrl string) ServiceClient {
-	s := ServiceRestClient{url: metaDbServiceUrl}
-
-	return &s
-}
-
-// Return an instance of DeviceProfileClient
-func NewDeviceProfileClient(metaDbDeviceProfileUrl string) DeviceProfileClient {
-	d := DeviceProfileRestClient{url: metaDbDeviceProfileUrl}
-
-	return &d
-}
-
-// Helper method to make the request and return the response
-func makeRequest(req *http.Request) (*http.Response, error) {
-	client := &http.Client{}
-	resp, err := client.Do(req)
-	return resp, err
-}
-
-// Helper method to get the body from the response after making the request
-func getBody(resp *http.Response) ([]byte, error) {
-	body, err := ioutil.ReadAll(resp.Body)
-
-	return body, err
-}
-
-// ***************** ADDRESSABLE CLIENT METHODS ***********************
-
-// Add an addressable - handle error codes
-// Returns the ID of the addressable and an error
-func (a *AddressableRestClient) Add(addr *models.Addressable) (string, error) {
-	// Marshal the addressable to JSON
-	jsonStr, err := json.Marshal(addr)
-	if err != nil {
-		return "", err
-	}
-
-	client := &http.Client{}
-	resp, err := client.Post(a.url, "application/json", bytes.NewReader(jsonStr))
-	if err != nil {
-		return "", err
-	}
-	if resp == nil {
-		return "", ErrResponseNil
-	}
-	defer resp.Body.Close()
-
-	// Get the response body
-	bodyBytes, err := getBody(resp)
-	if err != nil {
-		return "", err
-	}
-	bodyString := string(bodyBytes)
-
-	if resp.StatusCode != 200 {
-		return "", errors.New(string(bodyString))
-	}
-
-	return bodyString, err
-}
-
-// Helper method to decode an addressable and return the addressable
-func (d *AddressableRestClient) decodeAddressable(resp *http.Response) (models.Addressable, error) {
-	dec := json.NewDecoder(resp.Body)
-	addr := models.Addressable{}
-
-	err := dec.Decode(&addr)
-	if err != nil {
-		return models.Addressable{}, err
-	}
-
-	return addr, err
-}
-
-// TODO: make method signatures consistent wrt to error return value
-// ie. use it everywhere, or not at all!
-
-// Get the addressable by name
-func (a *AddressableRestClient) AddressableForName(name string) (models.Addressable, error) {
-	req, err := http.NewRequest(http.MethodGet, a.url+"/name/"+url.QueryEscape(name), nil)
-	if err != nil {
-		return models.Addressable{}, err
-	}
-
-	resp, err := makeRequest(req)
-
-	// Check response
-	if resp == nil {
-		return models.Addressable{}, ErrResponseNil
-	}
-	defer resp.Body.Close()
-	if err != nil {
-		fmt.Printf("AddressableForName makeRequest failed: %v\n", err)
-		return models.Addressable{}, err
-	}
-
-	if resp.StatusCode != 200 {
-		// Get the response body
-		bodyBytes, err := getBody(resp)
-		if err != nil {
-			return models.Addressable{}, err
-		}
-		bodyString := string(bodyBytes)
-
-		return models.Addressable{}, errors.New(bodyString)
-	}
-
-	return a.decodeAddressable(resp)
-}
-
-// ***************** DEVICE CLIENT METHODS ***********************
 // Help method to decode a device slice
 func (d *DeviceRestClient) decodeDeviceSlice(resp *http.Response) ([]models.Device, error) {
 	dec := json.NewDecoder(resp.Body)
@@ -1015,389 +825,4 @@ func (d *DeviceRestClient) DeleteByName(name string) error {
 	}
 
 	return nil
-}
-
-// ************************** COMMAND CLIENT METHODS ****************************
-
-// Helper method to decode and return a command
-func (c *CommandRestClient) decodeCommand(resp *http.Response) (models.Command, error) {
-	dec := json.NewDecoder(resp.Body)
-	com := models.Command{}
-	err := dec.Decode(&com)
-	if err != nil {
-		return models.Command{}, err
-	}
-
-	return com, err
-}
-
-// Helper method to decode and return a command slice
-func (c *CommandRestClient) decodeCommandSlice(resp *http.Response) ([]models.Command, error) {
-	dec := json.NewDecoder(resp.Body)
-	comSlice := []models.Command{}
-	err := dec.Decode(&comSlice)
-	if err != nil {
-		return []models.Command{}, err
-	}
-
-	return comSlice, err
-}
-
-// Get a command by id
-func (c *CommandRestClient) Command(id string) (models.Command, error) {
-	req, err := http.NewRequest(http.MethodGet, c.url+"/"+id, nil)
-	if err != nil {
-		return models.Command{}, err
-	}
-
-	resp, err := makeRequest(req)
-	if err != nil {
-		return models.Command{}, err
-	}
-	if resp == nil {
-		return models.Command{}, ErrResponseNil
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != 200 {
-		// Get the response body
-		bodyBytes, err := getBody(resp)
-		if err != nil {
-			return models.Command{}, err
-		}
-		bodyString := string(bodyBytes)
-
-		return models.Command{}, errors.New(bodyString)
-	}
-
-	return c.decodeCommand(resp)
-}
-
-// Get a list of all the commands
-func (c *CommandRestClient) Commands() ([]models.Command, error) {
-	req, err := http.NewRequest(http.MethodGet, c.url, nil)
-	if err != nil {
-		return []models.Command{}, err
-	}
-
-	resp, err := makeRequest(req)
-	if err != nil {
-		return []models.Command{}, err
-	}
-	if resp == nil {
-		return []models.Command{}, ErrResponseNil
-	}
-
-	if resp.StatusCode != 200 {
-		// Get the response body
-		bodyBytes, err := getBody(resp)
-		if err != nil {
-			return []models.Command{}, err
-		}
-		bodyString := string(bodyBytes)
-
-		return []models.Command{}, errors.New(bodyString)
-	}
-
-	return c.decodeCommandSlice(resp)
-}
-
-// Get a list of commands for a certain name
-func (c *CommandRestClient) CommandsForName(name string) ([]models.Command, error) {
-	req, err := http.NewRequest(http.MethodGet, c.url+"/name/"+name, nil)
-	if err != nil {
-		return []models.Command{}, err
-	}
-
-	resp, err := makeRequest(req)
-	if err != nil {
-		return []models.Command{}, err
-	}
-	if resp == nil {
-		return []models.Command{}, ErrResponseNil
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != 200 {
-		// Get the response body
-		bodyBytes, err := getBody(resp)
-		if err != nil {
-			return []models.Command{}, err
-		}
-		bodyString := string(bodyBytes)
-
-		return []models.Command{}, errors.New(bodyString)
-	}
-
-	return c.decodeCommandSlice(resp)
-}
-
-// Add a new command
-func (c *CommandRestClient) Add(com *models.Command) (string, error) {
-	jsonStr, err := json.Marshal(com)
-	if err != nil {
-		return "", err
-	}
-
-	req, err := http.NewRequest(http.MethodPost, c.url, bytes.NewReader(jsonStr))
-	if err != nil {
-		return "", err
-	}
-
-	resp, err := makeRequest(req)
-	if err != nil {
-		return "", err
-	}
-	if resp == nil {
-		return "", ErrResponseNil
-	}
-	defer resp.Body.Close()
-
-	// Get the response body
-	bodyBytes, err := getBody(resp)
-	if err != nil {
-		return "", err
-	}
-	bodyString := string(bodyBytes)
-
-	if resp.StatusCode != 200 {
-		return "", errors.New(bodyString)
-	}
-
-	return bodyString, nil
-}
-
-// Update a command
-func (c *CommandRestClient) Update(com models.Command) error {
-	jsonStr, err := json.Marshal(&com)
-	if err != nil {
-		return err
-	}
-
-	req, err := http.NewRequest(http.MethodPut, c.url, bytes.NewReader(jsonStr))
-	if err != nil {
-		return err
-	}
-
-	resp, err := makeRequest(req)
-	if err != nil {
-		return err
-	}
-	if resp == nil {
-		return ErrResponseNil
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != 200 {
-		// Get the response body
-		bodyBytes, err := getBody(resp)
-		if err != nil {
-			return err
-		}
-		bodyString := string(bodyBytes)
-
-		return errors.New(bodyString)
-	}
-
-	return nil
-}
-
-// Delete a command
-func (c *CommandRestClient) Delete(id string) error {
-	req, err := http.NewRequest(http.MethodDelete, c.url+"/id/"+id, nil)
-	if err != nil {
-		return err
-	}
-
-	resp, err := makeRequest(req)
-	if err != nil {
-		return err
-	}
-	if resp == nil {
-		return ErrResponseNil
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != 200 {
-		// Get the response body
-		bodyBytes, err := getBody(resp)
-		if err != nil {
-			return err
-		}
-		bodyString := string(bodyBytes)
-
-		return errors.New(bodyString)
-	}
-
-	return nil
-}
-
-// ********************** SERVICE CLIENT METHODS **************************
-// Helper method to decode and return a deviceservice
-func (s *ServiceRestClient) decodeDeviceService(resp *http.Response) (models.DeviceService, error) {
-	dec := json.NewDecoder(resp.Body)
-	ds := models.DeviceService{}
-	err := dec.Decode(&ds)
-	if err != nil {
-		return models.DeviceService{}, err
-	}
-
-	return ds, err
-}
-
-// Update the last connected time for the device service
-func (s *ServiceRestClient) UpdateLastConnected(id string, time int64) error {
-	req, err := http.NewRequest(http.MethodPut, s.url+"/"+id+"/lastconnected/"+strconv.FormatInt(time, 10), nil)
-	if err != nil {
-		return err
-	}
-
-	resp, err := makeRequest(req)
-	if err != nil {
-		return err
-	}
-	if resp == nil {
-		return ErrResponseNil
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != 200 {
-		// Get the response body
-		bodyBytes, err := getBody(resp)
-		if err != nil {
-			return err
-		}
-		bodyString := string(bodyBytes)
-
-		return errors.New(bodyString)
-	}
-
-	return nil
-}
-
-// Update the last reported time for the device service
-func (s *ServiceRestClient) UpdateLastReported(id string, time int64) error {
-	req, err := http.NewRequest(http.MethodPut, s.url+"/"+id+"/lastreported/"+strconv.FormatInt(time, 10), nil)
-	if err != nil {
-		return err
-	}
-
-	resp, err := makeRequest(req)
-	if err != nil {
-		return err
-	}
-	if resp == nil {
-		return ErrResponseNil
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != 200 {
-		// Get the response body
-		bodyBytes, err := getBody(resp)
-		if err != nil {
-			return err
-		}
-		bodyString := string(bodyBytes)
-
-		return errors.New(bodyString)
-	}
-
-	return nil
-}
-
-// Add a new deviceservice
-func (s *ServiceRestClient) Add(ds *models.DeviceService) (string, error) {
-	jsonStr, err := json.Marshal(ds)
-	if err != nil {
-		return "", err
-	}
-
-	client := &http.Client{}
-	resp, err := client.Post(s.url, "application/json", bytes.NewReader(jsonStr))
-	if err != nil {
-		return "", err
-	}
-	if resp == nil {
-		return "", ErrResponseNil
-	}
-	defer resp.Body.Close()
-
-	// Get the response body
-	bodyBytes, err := getBody(resp)
-	if err != nil {
-		return "", err
-	}
-	bodyString := string(bodyBytes)
-
-	if resp.StatusCode != 200 {
-		return "", errors.New(bodyString)
-	}
-
-	return bodyString, nil
-}
-
-// Request deviceservice for specified name
-func (s *ServiceRestClient) DeviceServiceForName(name string) (models.DeviceService, error) {
-	req, err := http.NewRequest(http.MethodGet, s.url+"/name/"+name, nil)
-	if err != nil {
-		fmt.Printf("DeviceServiceForName NewRequest failed: %v\n", err)
-		return models.DeviceService{}, err
-	}
-
-	resp, err := makeRequest(req)
-	if resp == nil {
-		return models.DeviceService{}, ErrResponseNil
-	}
-	defer resp.Body.Close()
-	if err != nil {
-		fmt.Printf("DeviceServiceForName makeRequest failed: %v\n", err)
-		return models.DeviceService{}, err
-	}
-
-	if resp.StatusCode != 200 {
-		// Get the response body
-		bodyBytes, err := getBody(resp)
-		if err != nil {
-			return models.DeviceService{}, err
-		}
-		bodyString := string(bodyBytes)
-
-		return models.DeviceService{}, errors.New(bodyString)
-	}
-
-	return s.decodeDeviceService(resp)
-}
-
-// ***************** DEVICE PROFILE METHODS *************************
-
-// Add a new device profile to metadata
-func (dpc *DeviceProfileRestClient) Add(dp *models.DeviceProfile) (string, error) {
-	jsonStr, err := json.Marshal(dp)
-	if err != nil {
-		return "", err
-	}
-
-	client := &http.Client{}
-	resp, err := client.Post(dpc.url, "application/json", bytes.NewReader(jsonStr))
-	if err != nil {
-		return "", err
-	}
-	if resp == nil {
-		return "", ErrResponseNil
-	}
-	defer resp.Body.Close()
-
-	// Get the response
-	bodyBytes, err := getBody(resp)
-	if err != nil {
-		return "", err
-	}
-	bodyString := string(bodyBytes)
-
-	// Check the response code
-	if resp.StatusCode != 200 {
-		return "", errors.New(bodyString)
-	}
-
-	return bodyString, nil
 }
