+++ core/clients/metadata/metadata.go
@@ -1,0 +1,40 @@
+/*******************************************************************************
+ * Copyright 2018 Dell Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License
+ * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing permissions and limitations under
+ * the License.
+ *******************************************************************************/
+package metadata
+
+import (
+	"errors"
+	"io/ioutil"
+	"net/http"
+)
+
+var (
+	ErrResponseNil error = errors.New("Problem connecting to metadata - reponse was nil")
+	ErrNotFound    error = errors.New("Item not found")
+)
+
+// Helper method to make the request and return the response
+func makeRequest(req *http.Request) (*http.Response, error) {
+	client := &http.Client{}
+	resp, err := client.Do(req)
+	return resp, err
+}
+
+// Helper method to get the body from the response after making the request
+func getBody(resp *http.Response) ([]byte, error) {
+	body, err := ioutil.ReadAll(resp.Body)
+
+	return body, err
+}
+
